third_party/blink/web_tests/fast/dom/shadow/gc-shadow-expected.txt

This tests that the JavaScript wrapper objects of shadow DOM objects
are not prematurely garbage collected.


PASS internals.shadowRoot(ps[0]).tattoo is "I <3 WebKit"
PASS internals.shadowRoot(ps[0]).styleSheets.tattoo is "I <3 WebKit cssom"
PASS internals.shadowRoot(ps[1]).tattoo is "I <3 WebKit"
PASS internals.shadowRoot(ps[1]).styleSheets.tattoo is "I <3 WebKit cssom"
PASS internals.shadowRoot(ps[2]).tattoo is "I <3 WebKit"
PASS internals.shadowRoot(ps[2]).styleSheets.tattoo is "I <3 WebKit cssom"
PASS internals.shadowRoot(ps[3]).tattoo is "I <3 WebKit"
PASS internals.shadowRoot(ps[3]).styleSheets.tattoo is "I <3 WebKit cssom"
PASS internals.shadowRoot(ps[4]).tattoo is "I <3 WebKit"
PASS internals.shadowRoot(ps[4]).styleSheets.tattoo is "I <3 WebKit cssom"
PASS internals.shadowRoot(ps[5]).tattoo is "I <3 WebKit"
PASS internals.shadowRoot(ps[5]).styleSheets.tattoo is "I <3 WebKit cssom"
PASS internals.shadowRoot(ps[6]).tattoo is "I <3 WebKit"
PASS internals.shadowRoot(ps[6]).styleSheets.tattoo is "I <3 WebKit cssom"
PASS internals.shadowRoot(ps[7]).tattoo is "I <3 WebKit"
PASS internals.shadowRoot(ps[7]).styleSheets.tattoo is "I <3 WebKit cssom"
PASS internals.shadowRoot(ps[8]).tattoo is "I <3 WebKit"
PASS internals.shadowRoot(ps[8]).styleSheets.tattoo is "I <3 WebKit cssom"
PASS internals.shadowRoot(ps[9]).tattoo is "I <3 WebKit"
PASS internals.shadowRoot(ps[9]).styleSheets.tattoo is "I <3 WebKit cssom"
PASS successfullyParsed is true

TEST COMPLETE
